requirements.txt
setuptools==69.5.1
pandas==2.2.2
colorama==0.4.6
torch_geometric==2.5.3
accelerate==0.30.1
numpy==1.26.4
wandb==0.17.7
colorlog==6.8.2
deepspeed
lightning
pytz
PyYAML
tensorboardX
tensorflow_cpu
tqdm
transformers
fbgemm_gpu
flash_attn
sentencepiece